Add silent and refactor
