+++ src/Esquio.UI/ClientApp/src/app/shared/auth/index.ts
@@ -1,3 +1,4 @@
 export * from './iauth.service';
 export * from './auth.service';
 export * from './user.model';
+export * from './auth.routes';
